+++ base/src/AzAudio/dsp/plugins/azaDSPDebugger.c
@@ -24,8 +24,8 @@ const azaDSP azaDSPDebuggerHeader = {
 	.guiMetadata = {
 		.name             = "DSP Debugger",
 		.selected         = 0,
-		.drawTargetWidth  = 0,
-		.drawCurrentWidth = 0,
+		.drawTargetWidth  = 0.0f,
+		.drawCurrentWidth = 0.0f,
 	},
 	.funcs = {
 		.fp_getSpecs = azaDSPDebuggerGetSpecs,
@@ -104,8 +104,8 @@ azaDSPSpecs azaDSPDebuggerGetSpecs(void *dsp, uint32_t samplerate) {
 
 void azagDrawDSPDebugger(void *dsp, azagRect bounds) {
 	azaDSPDebugger *data = dsp;
-	int boundsStartX = bounds.x;
-	int usedWidth;
+	float boundsStartX = bounds.x;
+	float usedWidth;
 	float latencyFrames = (float)data->config.specsToReport.latencyFrames;
 	usedWidth = azagDrawSliderFloat(bounds, &latencyFrames, 0.0f, 48000.0f, 1.0f, 0.0f, "Latency Frames", "%.0f");
 	data->config.specsToReport.latencyFrames = (uint32_t)roundf(latencyFrames);
@@ -120,6 +120,6 @@ void azagDrawDSPDebugger(void *dsp, azagRect bounds) {
 	usedWidth = azagDrawSliderFloat(bounds, &trailingFrames, 0.0f, 48000.0f, 1.0f, 0.0f, "Trailing Frames", "%.0f");
 	data->config.specsToReport.trailingFrames = (uint32_t)roundf(trailingFrames);
 	azagRectShrinkLeftMargin(&bounds, usedWidth);
-	int totalWidth = bounds.x - boundsStartX + azagThemeCurrent.margin.x;
+	float totalWidth = bounds.x - boundsStartX + azagThemeCurrent.margin.x;
 	data->dsp.guiMetadata.drawTargetWidth = totalWidth;
 }
